Fix three clang-tidy checks and clangd unused includes
readability-math-missing-parentheses,
readability-redundant-declaration and
bugprone-implicit-widening-of-multiplication-result are now clean in
src/.

Widening: i64 size constants use LL literals; pointer offsets cast to
ptrdiff_t; two loop counters are size_t.

EngineMupdfBegin/EndOperation were declared in both EngineAll.h and
EngineMupdf.h, and each copy was needed because the headers don't
include each other. ScopedEngineOperation's ctor/dtor moved to
EngineMupdf.cpp so EngineMupdf.h no longer needs them.

clangd only checks headers with include guards, so it flagged just
three includes. ImageReader.cpp's mupdf/fitz.h was redundant
(color-imp.h includes it). FilterBase.h and DarkModeSubclass.h are
needed (by PdfFilter.h and for the DarkMode alias) and are marked
IWYU pragma: keep.

prompt: fix clang-tidy readability-math-missing-parentheses, readability-redundant-declaration, bugprone-implicit-widening-of-multiplication-result and clangd unused-includes

+++ src/AnnotEditToolbar.cpp
@@ -910,7 +910,7 @@ static Pixmap* PixmapFromAnnotIconFz(fz_pixmap* src) {
     int srcAlpha = src->alpha;
     for (int y = 0; y < src->h; y++) {
         u8* s = src->samples + (src->stride * y);
-        u8* d = px->data + (px->stride * y);
+        u8* d = px->data + ((ptrdiff_t)px->stride * y);
         for (int x = 0; x < src->w; x++) {
             d[0] = s[2];
             d[1] = s[1];
@@ -1256,7 +1256,7 @@ static HBITMAP CreateMenuGlyphBitmap(int dx, int dy, PopupGlyphKind glyph, Str i
                 if (dyOut < 0 || dyOut >= dy) {
                     continue;
                 }
-                u8* s = src->data + (y * src->stride);
+                u8* s = src->data + ((ptrdiff_t)y * src->stride);
                 for (int x = 0; x < src->width; x++) {
                     int dxOut = x + ox;
                     if (dxOut < 0 || dxOut >= dx) {
